# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "接口文档",
        "title": "go-server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/private/v1/orders/lists": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "订单列表-分页",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderdto.PageResponse"
                        }
                    }
                }
            }
        },
        "/api/private/v1/orders/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "获取详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/orderdto.PublicDTO"
                        }
                    }
                }
            }
        },
...
        "/api/private/v1/roles": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "角色列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/roledto.RolePublicDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "创建",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/roledto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/roledto.RolePublicDTO"
                        }
                    }
                }
            }
        },
        "/api/private/v1/roles/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "更新当前用户",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/roledto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/roledto.RolePublicDTO"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.IsSend": {
            "type": "string",
            "enum": [
                "是",
                "否"
            ],
            "x-enum-varnames": [
                "IsSendYes",
                "IsSendNo"
            ]
        },
        "model.OrderPay": {
            "type": "string",
            "enum": [
                "0",
                "1",
                "2",
                "3"
            ],
            "x-enum-varnames": [
                "OrderPayUnpaid",
                "OrderPayAlipay",
                "OrderPayWechat",
                "OrderPayBankCard"
            ]
        },
        "model.PayStatus": {
            "type": "string",
            "enum": [
                "0",
                "1"
            ],
            "x-enum-varnames": [
                "PayStatusUnpaid",
                "PayStatusPaid"
            ]
        },
        "orderdto.PageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/orderdto.PublicDTO"
                    }
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "orderdto.PublicDTO": {
            "type": "object",
            "properties": {
                "consignee_addr": {
                    "description": "收货地址",
                    "type": "string"
                },
                "create_time": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "is_send": {
                    "description": "是否发货",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.IsSend"
                        }
                    ]
                },
                "order_fapiao_company": {
                    "description": "发票公司",
                    "type": "string"
                },
                "order_fapiao_content": {
                    "description": "发票内容",
                    "type": "string"
                },
                "order_fapiao_title": {
                    "description": "发票抬头",
                    "type": "string"
                },
                "order_id": {
                    "description": "订单id",
                    "type": "integer"
                },
                "order_number": {
                    "description": "订单号",
                    "type": "string"
                },
                "order_pay": {
                    "description": "支付方式",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.OrderPay"
                        }
                    ]
                },
                "order_price": {
                    "description": "订单价格",
                    "type": "number"
                },
                "pay_status": {
                    "description": "支付状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.PayStatus"
                        }
                    ]
                },
                "trade_no": {
                    "description": "交易号",
                    "type": "string"
                },
                "update_time": {
                    "description": "更新时间",
                    "type": "integer"
                },
                "user_id": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
        "roledto.LoginRequest": {
            "type": "object",
            "required": [
                "roleDesc",
                "roleName"
            ],
            "properties": {
                "roleDesc": {
                    "type": "string"
                },
                "roleName": {
                    "type": "string"
                }
            }
        },
        "roledto.RolePublicDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "角色ID",
                    "type": "integer"
                },
                "roleDesc": {
                    "description": "角色描述",
                    "type": "string"
                },
                "roleName": {
                    "description": "角色名称",
                    "type": "string"
                }
            }
        }
    }
}